Versión estable con nuevo servidor.

+++ dev/laravel-backend/database/migrations/2018_01_15_050141_create_categories_products_table.php
@@ -13,12 +13,11 @@ class CreateCategoriesProductsTable extends Migration
      */
     public function up()
     {
+        if (Schema::hasTable('categories_products')) return;
         Schema::create('categories_products', function (Blueprint $table) {
             $table->increments('caterorie_id');
             $table->unsignedInteger('product_id');
-
-            $table->index(["caterorie_id"], 'fk_categori_product_categories1_idx');
-
+            $table->index(["categorie_id"], 'fk_categori_product_categories1_idx');
             $table->foreign('product_id', 'fk_categori_product_products1_idx')
                 ->references('id')->on('product')
                 ->onDelete('cascade')
